moved props back inside component
